Makes browse cards more uniform.
Based on communication with evanwill on CollectionBuilder Slack:

There are a couple ways you could-- the Browse cards have an image cap and a text card body. In the default set up, neither is a set height. The image size is constrained by the width (a col-lg-4 which will be around 400px on a large screen), but the height is what ever the natural ratio of the image is. The size of the card body will be different based on how much metadata the item has. So if you wanted the images to all be the same height (but not distort the ratio), you could add some CSS like:
.card-img-top { height: 200px; object-fit: cover; }
see the options for object-fit, https://developer.mozilla.org/en-US/docs/Web/CSS/object-fit
The card body is more difficult to make consistent size. The best option is probably to add h-100 to the card. You would have to edit "_includes/js/browse-js.html" line 42 like,
2:30
var card = '<div class="item col-lg-4 col-md-6 mb-2"><div class="card h-100">';